Create generateCppClass.sh

+++ .aux/etc/generate_c_structure.sh
@@ -12,8 +12,13 @@ else
 	exit 1
 fi
 
+if [[ -e $2/$1.h ]] || [[ -e $2/$1.c ]]; then
+	echo "[$2/$1.h] or [$2/$1.c] is already exists!"
+	exit 1
+fi
+
 cd $env/.tool
-echo uppercase name = $(echo $1 | tr "[:lower:]" "[:upper:]")"\n"lowercase name = $1 > name.variable
+echo uppercase name = $(tr '[:lower:]' '[:upper:]' <<< $1)"\n"lowercase name = $1 > name.variable
 ./variabledText.out template.h -o $2/$1.h
 ./variabledText.out template.c -o $2/$1.c
 rm name.variable
